automatically remove group if last file in it is closed

+++ src/mainwindow.cpp
@@ -1692,8 +1692,14 @@ void MainWindow::closeFile(FileContext* file)
 {
     if (!file->isModified() || requestCloseChanged(QList<FileContext*>() << file)) {
         ui->projectView->setCurrentIndex(QModelIndex());
-        fileClosed(file->id());
-        mFileRepo.removeFile(file);
+
+        if (file->parentEntry()->childCount() > 1) {
+            fileClosed(file->id());
+            mFileRepo.removeFile(file);
+        } else {
+            closeGroup(file->parentEntry());
+        }
+
         mSettings->saveSettings(this);
     }
 }
